requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
bleach==6.0.0
cachetools==5.3.0
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.1
cloudpickle==2.2.1
cmake==3.26.1
cohere==4.3.1
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
defusedxml==0.7.1
dm-tree==0.1.8
docopt==0.6.2
environs==9.5.0
etils==1.1.1
executing==1.2.0
fastapi==0.95.1
fastjsonschema==2.16.3
filelock==3.10.2
flatbuffers==2.0.7
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
funcsigs==1.0.2
gast==0.4.0
gin-config==0.5.0
google-auth==2.17.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
greenlet==1.1.2
grpcio==1.63.0
grpcio-tools==1.63.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.8.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.13.3
humanfriendly==10.0
idna==3.7
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jax==0.4.8
jaxlib==0.4.7
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-server==1.23.6
jupyter-ydoc==0.2.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.0
kaggle==1.5.13
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lit==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.21.2
matplotlib==3.3.4
matplotlib-inline==0.1.6
milvus_cli==0.4.2
minio==7.2.7
mistune==2.0.5
ml-dtypes==0.0.4
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.4
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.0.26
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvcc-cu12==12.1.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.55
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.8.1.3
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.4
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.4.55
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.0.2.55
nvidia-nccl-cu11==2.14.3
nvidia-nvjitlink-cu12==12.1.55
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.14.0
onnxruntime==1.14.1
openai==0.27.6
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.11
-e git+ssh://git@github.com/LashaJini/mind-palace.git@593145b56df10e0c19d3e615566f9e44f6009277#egg=pkg
platformdirs==3.2.0
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==5.26.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.7
PyJWT==2.6.0
pymilvus==2.3.4
pynvim==0.4.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3
pyzmq==25.0.2
redis==4.5.4
regex==2023.3.23
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
sympy==1.11.1
tabulate==0.8.9
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-datasets==4.8.3
tensorflow-estimator==2.7.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.12.0
tensorflow-text==2.12.0
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers@b61d5b47f640308068139561f673765b2af39874
trax==1.4.1
triton==2.0.0
typing_extensions==4.5.0
tzdata==2024.1
ujson==5.10.0
uri-template==1.2.0
urllib3==2.2.1
uvicorn==0.22.0
wcwidth==0.2.6
webcolors==1.13
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.14.1
y-py==0.5.9
yarg==0.1.9
yarl==1.9.2
ypy-websocket==0.8.2
zipp==3.15.0
